+++ RELEASE_NOTES.md
@@ -8,6 +8,9 @@ _Date TBD_
 * Updated /certification_ids/ call. Still returns the list of certification_ids and the date each was created. If user is ROLE_ADMIN, also returns a semi-colon separated list of CHPL product numbers with each certification_id.
 * Remove terms of use and api documentation from certified products.
 * Check product code, version code, and ICS code against specific sets of characters to make sure no special characters are used.
+* Created Java program to send weekly email with aggregate counts for developers, products, certified products, and 2014 + 2015 CPs
+	* weekly email will be sent to Jennifer Hui who will package it with the weekly status report
+	* [TEMPLATE] for deployment, update the properties.environment summaryEmail property to send to jhui@ainq.com
 
 ### Bugs Fixed
 
@@ -46,6 +49,7 @@ _30 August 2016_
 * TO DO DURING THE RELEASE: change the questionable activity email recipients to just be the ONC_CHPL@hhs.gov email (i.e. remove onc.certification@hhs.gov)
 * TO DO DURING THE RELEASE: create a file (can call it cleantomcat) in /etc/cron.daily and chmod a+x the file. Contents of the file are listed as a comment in OCD-811. The command deletes files that have not been written to since X days ago. The number near the end of the command is X.
 * Updated /activity API endpoint to incorporate new parameters to filter by API-Key, sort dateAscending, and filter by start & end date.
+
 * Do not allow 170.315 (d)(3) to mark GAP as true
 * Added 'responsibleUser' field with all user data for /activity reports
 * Removed CORSFilter in web.xml; this was preventing some ajax calls from other domains